Move type checking util functions to custom_types package
